+++ crates/wdl-engine/src/stdlib.rs
@@ -1,26 +1,37 @@
 //! Module for the WDL standard library implementation.
 
+use std::borrow::Cow;
 use std::collections::HashMap;
 use std::sync::LazyLock;
 
+use regex::Regex;
 use wdl_analysis::stdlib::Binding;
+use wdl_analysis::types::Optional;
 use wdl_analysis::types::PrimitiveTypeKind;
 use wdl_analysis::types::Type;
 use wdl_analysis::types::TypeEq;
 use wdl_analysis::types::Types;
 use wdl_ast::Diagnostic;
+use wdl_ast::Span;
 
 use crate::Coercible;
+use crate::PrimitiveValue;
 use crate::Value;
+use crate::diagnostics::invalid_regex;
 
 /// Rounds a floating point number down to the next lower integer.
 ///
 /// https://github.com/openwdl/wdl/blob/wdl-1.2/SPEC.md#floor
-pub fn floor(types: &Types, arguments: &[Value], return_type: Type) -> Result<Value, Diagnostic> {
+pub fn floor(
+    types: &Types,
+    arguments: &[(Value, Span)],
+    return_type: Type,
+) -> Result<Value, Diagnostic> {
     debug_assert_eq!(arguments.len(), 1);
     debug_assert!(return_type.type_eq(types, &PrimitiveTypeKind::Integer.into()));
 
     let arg = arguments[0]
+        .0
         .coerce(types, PrimitiveTypeKind::Float.into())
         .expect("value should coerce to float")
         .unwrap_float();
@@ -30,11 +41,16 @@ pub fn floor(types: &Types, arguments: &[Value], return_type: Type) -> Result<Va
 /// Rounds a floating point number up to the next higher integer.
 ///
 /// https://github.com/openwdl/wdl/blob/wdl-1.2/SPEC.md#ceil
-pub fn ceil(types: &Types, arguments: &[Value], return_type: Type) -> Result<Value, Diagnostic> {
+pub fn ceil(
+    types: &Types,
+    arguments: &[(Value, Span)],
+    return_type: Type,
+) -> Result<Value, Diagnostic> {
     debug_assert_eq!(arguments.len(), 1);
     debug_assert!(return_type.type_eq(types, &PrimitiveTypeKind::Integer.into()));
 
     let arg = arguments[0]
+        .0
         .coerce(types, PrimitiveTypeKind::Float.into())
         .expect("value should coerce to float")
         .unwrap_float();
@@ -45,11 +61,16 @@ pub fn ceil(types: &Types, arguments: &[Value], return_type: Type) -> Result<Val
 /// rounding rules ("round half up").
 ///
 /// https://github.com/openwdl/wdl/blob/wdl-1.2/SPEC.md#round
-pub fn round(types: &Types, arguments: &[Value], return_type: Type) -> Result<Value, Diagnostic> {
+pub fn round(
+    types: &Types,
+    arguments: &[(Value, Span)],
+    return_type: Type,
+) -> Result<Value, Diagnostic> {
     debug_assert_eq!(arguments.len(), 1);
     debug_assert!(return_type.type_eq(types, &PrimitiveTypeKind::Integer.into()));
 
     let arg = arguments[0]
+        .0
         .coerce(types, PrimitiveTypeKind::Float.into())
         .expect("value should coerce to float")
         .unwrap_float();
@@ -59,15 +80,21 @@ pub fn round(types: &Types, arguments: &[Value], return_type: Type) -> Result<Va
 /// Returns the smaller of two integer values.
 ///
 /// https://github.com/openwdl/wdl/blob/wdl-1.2/SPEC.md#min
-pub fn int_min(types: &Types, arguments: &[Value], return_type: Type) -> Result<Value, Diagnostic> {
+pub fn int_min(
+    types: &Types,
+    arguments: &[(Value, Span)],
+    return_type: Type,
+) -> Result<Value, Diagnostic> {
     debug_assert_eq!(arguments.len(), 2);
     debug_assert!(return_type.type_eq(types, &PrimitiveTypeKind::Integer.into()));
 
     let first = arguments[0]
+        .0
         .coerce(types, PrimitiveTypeKind::Integer.into())
         .expect("value should coerce to integer")
         .unwrap_integer();
     let second = arguments[1]
+        .0
         .coerce(types, PrimitiveTypeKind::Integer.into())
         .expect("value should coerce to integer")
         .unwrap_integer();
@@ -79,17 +106,19 @@ pub fn int_min(types: &Types, arguments: &[Value], return_type: Type) -> Result<
 /// https://github.com/openwdl/wdl/blob/wdl-1.2/SPEC.md#min
 pub fn float_min(
     types: &Types,
-    arguments: &[Value],
+    arguments: &[(Value, Span)],
     return_type: Type,
 ) -> Result<Value, Diagnostic> {
     debug_assert_eq!(arguments.len(), 2);
     debug_assert!(return_type.type_eq(types, &PrimitiveTypeKind::Float.into()));
 
     let first = arguments[0]
+        .0
         .coerce(types, PrimitiveTypeKind::Float.into())
         .expect("value should coerce to float")
         .unwrap_float();
     let second = arguments[1]
+        .0
         .coerce(types, PrimitiveTypeKind::Float.into())
         .expect("value should coerce to float")
         .unwrap_float();
@@ -99,15 +128,21 @@ pub fn float_min(
 /// Returns the larger of two integer values.
 ///
 /// https://github.com/openwdl/wdl/blob/wdl-1.2/SPEC.md#max
-pub fn int_max(types: &Types, arguments: &[Value], return_type: Type) -> Result<Value, Diagnostic> {
+pub fn int_max(
+    types: &Types,
+    arguments: &[(Value, Span)],
+    return_type: Type,
+) -> Result<Value, Diagnostic> {
     assert_eq!(arguments.len(), 2);
     debug_assert!(return_type.type_eq(types, &PrimitiveTypeKind::Integer.into()));
 
     let first = arguments[0]
+        .0
         .coerce(types, PrimitiveTypeKind::Integer.into())
         .expect("value should coerce to integer")
         .unwrap_integer();
     let second = arguments[1]
+        .0
         .coerce(types, PrimitiveTypeKind::Integer.into())
         .expect("value should coerce to integer")
         .unwrap_integer();
@@ -119,25 +154,128 @@ pub fn int_max(types: &Types, arguments: &[Value], return_type: Type) -> Result<
 /// https://github.com/openwdl/wdl/blob/wdl-1.2/SPEC.md#max
 pub fn float_max(
     types: &Types,
-    arguments: &[Value],
+    arguments: &[(Value, Span)],
     return_type: Type,
 ) -> Result<Value, Diagnostic> {
     assert_eq!(arguments.len(), 2);
     debug_assert!(return_type.type_eq(types, &PrimitiveTypeKind::Float.into()));
 
     let first = arguments[0]
+        .0
         .coerce(types, PrimitiveTypeKind::Float.into())
         .expect("value should coerce to float")
         .unwrap_float();
     let second = arguments[1]
+        .0
         .coerce(types, PrimitiveTypeKind::Float.into())
         .expect("value should coerce to float")
         .unwrap_float();
     Ok(first.max(second).into())
 }
 
+/// Given two String parameters `input` and `pattern`, searches for the
+/// occurrence of `pattern` within `input` and returns the first match or `None`
+/// if there are no matches.
+///
+/// https://github.com/openwdl/wdl/blob/wdl-1.2/SPEC.md#-find
+pub fn find(
+    types: &Types,
+    arguments: &[(Value, Span)],
+    return_type: Type,
+) -> Result<Value, Diagnostic> {
+    assert_eq!(arguments.len(), 2);
+    debug_assert!(return_type.type_eq(types, &Type::from(PrimitiveTypeKind::String).optional()));
+
+    let input = arguments[0]
+        .0
+        .coerce(types, PrimitiveTypeKind::String.into())
+        .expect("value should coerce to string")
+        .unwrap_string();
+    let pattern = arguments[1]
+        .0
+        .coerce(types, PrimitiveTypeKind::String.into())
+        .expect("value should coerce to string")
+        .unwrap_string();
+
+    let regex = Regex::new(pattern.as_str()).map_err(|e| invalid_regex(&e, arguments[1].1))?;
+
+    match regex.find(input.as_str()) {
+        Some(m) => Ok(PrimitiveValue::new_string(m.as_str()).into()),
+        None => Ok(Value::None),
+    }
+}
+
+/// Given two String parameters `input` and `pattern`, tests whether `pattern`
+/// matches `input` at least once.
+///
+/// https://github.com/openwdl/wdl/blob/wdl-1.2/SPEC.md#-matches
+pub fn matches(
+    types: &Types,
+    arguments: &[(Value, Span)],
+    return_type: Type,
+) -> Result<Value, Diagnostic> {
+    assert_eq!(arguments.len(), 2);
+    debug_assert!(return_type.type_eq(types, &PrimitiveTypeKind::Boolean.into()));
+
+    let input = arguments[0]
+        .0
+        .coerce(types, PrimitiveTypeKind::String.into())
+        .expect("value should coerce to string")
+        .unwrap_string();
+    let pattern = arguments[1]
+        .0
+        .coerce(types, PrimitiveTypeKind::String.into())
+        .expect("value should coerce to string")
+        .unwrap_string();
+
+    let regex = Regex::new(pattern.as_str()).map_err(|e| invalid_regex(&e, arguments[1].1))?;
+    Ok(regex.is_match(input.as_str()).into())
+}
+
+/// Given three String parameters `input`, `pattern`, and `replace`, this
+/// function replaces all non-overlapping occurrences of `pattern` in `input`
+/// with `replace`.
+///
+/// https://github.com/openwdl/wdl/blob/wdl-1.2/SPEC.md#sub
+pub fn sub(
+    types: &Types,
+    arguments: &[(Value, Span)],
+    return_type: Type,
+) -> Result<Value, Diagnostic> {
+    assert_eq!(arguments.len(), 3);
+    debug_assert!(return_type.type_eq(types, &PrimitiveTypeKind::String.into()));
+
+    let input = arguments[0]
+        .0
+        .coerce(types, PrimitiveTypeKind::String.into())
+        .expect("value should coerce to string")
+        .unwrap_string();
+    let pattern = arguments[1]
+        .0
+        .coerce(types, PrimitiveTypeKind::String.into())
+        .expect("value should coerce to string")
+        .unwrap_string();
+    let replacement = arguments[2]
+        .0
+        .coerce(types, PrimitiveTypeKind::String.into())
+        .expect("value should coerce to string")
+        .unwrap_string();
+
+    let regex = Regex::new(pattern.as_str()).map_err(|e| invalid_regex(&e, arguments[1].1))?;
+    match regex.replace(input.as_str(), replacement.as_str()) {
+        Cow::Borrowed(_) => {
+            // No replacements, just return the input
+            Ok(PrimitiveValue::String(input).into())
+        }
+        Cow::Owned(s) => {
+            // A replacement occurred, allocate a new string
+            Ok(PrimitiveValue::new_string(s).into())
+        }
+    }
+}
+
 /// Represents a WDL function implementation callback.
-type Callback = fn(&Types, &[Value], Type) -> Result<Value, Diagnostic>;
+type Callback = fn(&Types, &[(Value, Span)], Type) -> Result<Value, Diagnostic>;
 
 /// Represents an implementation signature for a WDL standard library function.
 #[derive(Debug, Clone, Copy)]
@@ -177,7 +315,7 @@ impl Function {
         &self,
         binding: Binding<'_>,
         types: &Types,
-        arguments: &[Value],
+        arguments: &[(Value, Span)],
     ) -> Result<Value, Diagnostic> {
         (self.signatures[binding.index()].callback)(types, arguments, binding.return_type())
     }
@@ -262,6 +400,32 @@ pub static STDLIB: LazyLock<StandardLibrary> = LazyLock::new(|| {
             )
             .is_none()
     );
+    assert!(
+        functions
+            .insert(
+                "find",
+                Function::new(const { &[Signature::new("(String, String) -> String?", find)] })
+            )
+            .is_none()
+    );
+    assert!(
+        functions
+            .insert(
+                "matches",
+                Function::new(const { &[Signature::new("(String, String) -> Boolean", matches)] })
+            )
+            .is_none()
+    );
+    assert!(
+        functions
+            .insert(
+                "sub",
+                Function::new(
+                    const { &[Signature::new("(String, String, String) -> String", sub,)] }
+                )
+            )
+            .is_none()
+    );
 
     StandardLibrary { functions }
 });
@@ -285,7 +449,11 @@ mod test {
             match STDLIB.functions.get(name) {
                 Some(imp) => match func {
                     wdl_analysis::stdlib::Function::Monomorphic(f) => {
-                        assert_eq!(imp.signatures.len(), 1);
+                        assert_eq!(
+                            imp.signatures.len(),
+                            1,
+                            "signature mismatch for function `{name}`"
+                        );
                         assert_eq!(
                             f.signature()
                                 .display(
@@ -293,7 +461,8 @@ mod test {
                                     &TypeParameters::new(f.signature().type_parameters())
                                 )
                                 .to_string(),
-                            imp.signatures[0].display
+                            imp.signatures[0].display,
+                            "signature mismatch for function `{name}`"
                         );
                     }
                     wdl_analysis::stdlib::Function::Polymorphic(f) => {
@@ -305,7 +474,8 @@ mod test {
                                     &TypeParameters::new(sig.type_parameters())
                                 )
                                 .to_string(),
-                                imp.signatures[i].display
+                                imp.signatures[i].display,
+                                "signature mismatch for function `{name}` (index {i})"
                             );
                         }
                     }
@@ -543,5 +713,102 @@ mod test {
         )
         .unwrap();
         approx::assert_relative_eq!(value.unwrap_float(), 54321.0);
+    }
+
+    #[test]
+    fn find() {
+        let scopes = &[Scope::new(None)];
+        let scope = ScopeRef::new(scopes, 0);
+
+        let mut types = Types::default();
+        let diagnostic =
+            eval_v1_expr(V1::Two, "find('foo bar baz', '?')", &mut types, scope).unwrap_err();
+        assert_eq!(
+            diagnostic.message(),
+            "regex parse error:\n    ?\n    ^\nerror: repetition operator missing expression"
+        );
+
+        let value =
+            eval_v1_expr(V1::Two, "find('hello world', 'e..o')", &mut types, scope).unwrap();
+        assert_eq!(value.unwrap_string().as_str(), "ello");
+
+        let value =
+            eval_v1_expr(V1::Two, "find('hello world', 'goodbye')", &mut types, scope).unwrap();
+        assert!(value.is_none());
+
+        let value = eval_v1_expr(V1::Two, "find('hello\tBob', '\\t')", &mut types, scope).unwrap();
+        assert_eq!(value.unwrap_string().as_str(), "\t");
+    }
+
+    #[test]
+    fn matches() {
+        let scopes = &[Scope::new(None)];
+        let scope = ScopeRef::new(scopes, 0);
+
+        let mut types = Types::default();
+        let diagnostic =
+            eval_v1_expr(V1::Two, "matches('foo bar baz', '?')", &mut types, scope).unwrap_err();
+        assert_eq!(
+            diagnostic.message(),
+            "regex parse error:\n    ?\n    ^\nerror: repetition operator missing expression"
+        );
+
+        let value =
+            eval_v1_expr(V1::Two, "matches('hello world', 'e..o')", &mut types, scope).unwrap();
+        assert!(value.unwrap_boolean());
+
+        let value = eval_v1_expr(
+            V1::Two,
+            "matches('hello world', 'goodbye')",
+            &mut types,
+            scope,
+        )
+        .unwrap();
+        assert!(!value.unwrap_boolean());
+
+        let value =
+            eval_v1_expr(V1::Two, "matches('hello\tBob', '\\t')", &mut types, scope).unwrap();
+        assert!(value.unwrap_boolean());
+    }
+
+    #[test]
+    fn sub() {
+        let scopes = &[Scope::new(None)];
+        let scope = ScopeRef::new(scopes, 0);
+
+        let mut types = Types::default();
+        let diagnostic = eval_v1_expr(
+            V1::Two,
+            "sub('foo bar baz', '?', 'nope')",
+            &mut types,
+            scope,
+        )
+        .unwrap_err();
+        assert_eq!(
+            diagnostic.message(),
+            "regex parse error:\n    ?\n    ^\nerror: repetition operator missing expression"
+        );
+
+        let value = eval_v1_expr(
+            V1::Two,
+            "sub('hello world', 'e..o', 'ey there')",
+            &mut types,
+            scope,
+        )
+        .unwrap();
+        assert_eq!(value.unwrap_string().as_str(), "hey there world");
+
+        let value = eval_v1_expr(
+            V1::Two,
+            "sub('hello world', 'goodbye', 'nope')",
+            &mut types,
+            scope,
+        )
+        .unwrap();
+        assert_eq!(value.unwrap_string().as_str(), "hello world");
+
+        let value =
+            eval_v1_expr(V1::Two, "sub('hello\tBob', '\\t', ' ')", &mut types, scope).unwrap();
+        assert_eq!(value.unwrap_string().as_str(), "hello Bob");
     }
 }